+++ tests/test_bw_rank_est.cpp
@@ -39,7 +39,7 @@ void test_all_64_valid_uniform() {
     std::cout << "Testing all 64 valid columns, uniform degrees..." << std::endl;
     LingenResult R;
     R.valid_mask = ~uint64_t(0);  // all 64 bits set
-    for (int j = 0; j < 64; ++j) R.degrees[j] = 100;
+    for (int j = 0; j < 64; ++j) R.degrees[static_cast<size_t>(j)] = 100;
     assert(compute_rank_est(R) == 6400);
     std::cout << "  PASS" << std::endl;
 }
@@ -49,7 +49,7 @@ void test_mask_selects_subset() {
     LingenResult R;
     // bits 0, 5, 17, 63 set; only those degrees count
     R.valid_mask = (1ULL << 0) | (1ULL << 5) | (1ULL << 17) | (1ULL << 63);
-    for (int j = 0; j < 64; ++j) R.degrees[j] = j + 1;  // 1..64
+    for (int j = 0; j < 64; ++j) R.degrees[static_cast<size_t>(j)] = j + 1;  // 1..64
     // expected = degrees[0] + degrees[5] + degrees[17] + degrees[63]
     //         = 1 + 6 + 18 + 64 = 89
     assert(compute_rank_est(R) == 89);
@@ -61,7 +61,7 @@ void test_invalid_cols_ignored_even_with_large_degrees() {
     LingenResult R;
     R.valid_mask = (1ULL << 0);
     R.degrees[0] = 7;
-    for (int j = 1; j < 64; ++j) R.degrees[j] = 1000000;  // garbage in invalid slots
+    for (int j = 1; j < 64; ++j) R.degrees[static_cast<size_t>(j)] = 1000000;  // garbage in invalid slots
     assert(compute_rank_est(R) == 7);
     std::cout << "  PASS" << std::endl;
 }
@@ -72,7 +72,7 @@ void test_full_rank_thin_solve_like() {
     // Realistic profile: most cols valid, average degree ~85
     LingenResult R;
     R.valid_mask = ~uint64_t(0);
-    for (int j = 0; j < 64; ++j) R.degrees[j] = 80;
+    for (int j = 0; j < 64; ++j) R.degrees[static_cast<size_t>(j)] = 80;
     assert(compute_rank_est(R) == 64 * 80);  // 5120
     // This sits ≈ m=5200 → matrix has excess → BW should find deps
     std::cout << "  PASS" << std::endl;
